requirements.txt
paho-mqtt
influxdb
pytest
